Eliminate constant errors by checking for existence of source files.

+++ trunk/VirtualAGC/CopyApolloSource.sh
@@ -14,13 +14,26 @@ DESTDIR=$1/source/$SOURCENAME
 EXT=$3
 WIN=$4
 
+SRCPATHS="*.binsource *.agc *.aea Apollo32.png *.agc.html *.aea.html"
+
 mkdir $DESTDIR
+
 cp ../Contributed/SyntaxHighlight/Prettify/*.js $DESTDIR
 cp ../Contributed/SyntaxHighlight/Prettify/*.css $DESTDIR
-cp $EXT $SOURCEDIR/*.binsource $SOURCEDIR/*.agc $DESTDIR
-cp $EXT $SOURCEDIR/*.binsource $SOURCEDIR/*.aea $DESTDIR
-#sh ./lst2html.sh $SOURCEDIR/$SOURCENAME.lst $DESTDIR/$SOURCENAME.html $WIN
-cp $SOURCEDIR/*.agc.html $SOURCEDIR/*.aea.html $SOURCEDIR/Apollo32.png $DESTDIR
-cp $EXT $SOURCEDIR/$SOURCENAME.bin $DESTDIR
-cp $EXT $SOURCEDIR/$SOURCENAME.symtab $DESTDIR/$SOURCENAME.bin.symtab
 
+for srcpath in ${SRCPATHS}; do 
+    sources=`find $SOURCEDIR -name $srcpath`
+    if [ ! -z "$sources" ]; then
+        cp $EXT $sources $DESTDIR
+    fi
+done
+ 
+#sh ./lst2html.sh $SOURCEDIR/$SOURCENAME.lst $DESTDIR/$SOURCENAME.html $WIN
+
+if [ -f $SOURCEDIR/$SOURCENAME.bin ]; then
+    cp $EXT $SOURCEDIR/$SOURCENAME.bin $DESTDIR
+fi
+
+if [ -f $SOURCEDIR/$SOURCENAME.symtab ]; then
+    cp $EXT $SOURCEDIR/$SOURCENAME.symtab $DESTDIR/$SOURCENAME.bin.symtab
+fi
